[BO] Updated function on edit pages following changing type of Id (int to long)

+++ BackOffice/Pages/Brands/Edit.cshtml.cs
@@ -76,7 +76,7 @@ public class EditModel : PageModel
         return RedirectToPage("./Index");
     }
 
-    private bool BrandEntityExists(int id)
+    private bool BrandEntityExists(long id)
     {
         return _context.Brands.Any(e => e.Id == id);
     }
